Remove GoogleService-Info.plist from .gitignore
This file is usually included in the .gitignore to encourage others to setup their own firebase backend when pulling/forking a repo, as well as the database being in test mode. As a result, the test step of the GitHub actions build workflow was failing. The database has been moved into production mode with proper read/write rules, so this can now be safely removed from the .gitignore, allowing the test step in the build workflow to run.